pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argus-scanner"
version = "2.0.0"
description = "AI-powered open source vulnerability scanner. Built for defenders."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["security", "vulnerability", "sast", "scanner", "ai", "devsecops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "aiohttp>=3.9",
    "click>=8.0",
]

[project.optional-dependencies]
watch = ["watchdog>=4.0"]
yaml  = ["PyYAML>=6.0"]
full  = ["watchdog>=4.0", "PyYAML>=6.0"]

[project.scripts]
argus = "argus.cli:main"

[project.urls]
Homepage   = "https://github.com/stackbleed-ctrl/ARGUS"
Repository = "https://github.com/stackbleed-ctrl/ARGUS"
Issues     = "https://github.com/stackbleed-ctrl/ARGUS/issues"

[tool.hatch.build.targets.wheel]
packages = ["argus"]
